LPS-114893 Move closure package.
