+++ modules/apps/change-tracking/change-tracking-api/bnd.bnd
@@ -2,10 +2,10 @@ Bundle-Name: Liferay Change Tracking API
 Bundle-SymbolicName: com.liferay.change.tracking.api
 Bundle-Version: 11.0.0
 Export-Package:\
+	com.liferay.change.tracking.closure,\
 	com.liferay.change.tracking.conflict,\
 	com.liferay.change.tracking.constants,\
 	com.liferay.change.tracking.exception,\
 	com.liferay.change.tracking.model,\
-	com.liferay.change.tracking.reference.closure,\
 	com.liferay.change.tracking.service,\
 	com.liferay.change.tracking.service.persistence
